save to mysql is working.

+++ src/PitWallDataGatheringApi/PitWallDataGatheringApi.Repositories.Integration.Tests/Gauges/Sql/ActualLeaderboardRow.cs
@@ -10,7 +10,7 @@
         public string car_name { get; set; }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
 
-        public int data_tick { get; set; }
+        public long data_tick { get; set; }
 
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string metric_name { get; set; }
@@ -18,6 +18,7 @@
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string metric_value { get; set; }
 #pragma warning restore CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
     }
 
 }
